rpc: Calculate delay and export it via metrics
Now the client has two durations using which it can estimate the
rount-trip time of a packet -- the duration between sending request and
receiving response and the time it too handler to execute on the server
side. Difference is the delay value.

Those values from different calls are accumulated in a client counter as
well as the total number of delay samples received (errors and old
servers don't send it). Metrics export both as counters, so one can rate
it over time and monitor average delay.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>