test: Add test for sharded<>::invoke_on_...() compilation
There are several options what invoke_on_...() can try to invoke. So far
only non-const and non-mutable callables are in use and it works.

If calling it with mutable lambdas, it used to fail compilation, but
previous patch fixed it, so here's the test.

It also makes sense to support calling it on const sharded<> object
(with the lambda accepting const service reference), but it's not that
simple (see #2278)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>