json: Make formatter call write for jsonable
`write(output_stream<char>&, const jsonable&)` would call
`jsonable::to_json` which would just serialize the whole object (and its
children) to a string and then dump that onto the `output_stream`.

Instead call `jsonable::write(output_stream<char>&)` which will
recursively call `write` (for the `jsonbase` subclasses)  and hence
propagate the benefits of writing to the output stream directly.

To make sure that the object is alive for the duration of the call we
change the overload to copy its argument.