reactor: respect preempt requests in debug mode
In [e0d515b6 reactor: poll less frequently in debug mode][1] reactor
preemption logic in debug mode was changed to improve performance. It
also introduced a problem causing `seastar::lowres_clock` to drift
unbounded.

The code doc states that `sestar::lowres_clock` "is a monotonic clock
with a granularity of ~task_quota". We want that claim to be respected.

To achieve that we change the `scheduler_need_preempt` to respect the
preemption requests which are triggered e.g. when task quota is up.

[1]: https://github.com/scylladb/seastar/commit/e0d515b6cf96129cebdcbb1b031df90eb815e7fa

Closes https://github.com/scylladb/seastar/issues/2238