io_sink: use chunked_fifo as _pending_io container
This is to avoid the large contiguous allocation done by
circular_buffer when there many pending IOs, as such allocations can
fail on a long running system due to fragmentation.

Performance should be comparable to the existing approach: based on
microbenchmarks, iteration and erasing elements is similar with
chunked_fifo, perhaps 10% to 20% slower for those operations on modestly
sized containers.

Fixes #2355.