abort_source: subscription: keep callback function alive after abort
Change 470b539b1c1dd2b080eb13cc77b17eec5a83c21f causes
a regression in scylla when the subscription callback
function is destroyed after being called when abort
is requested.

The app expects this function to remain alive throughout
the life time of the subscription object holding it,
so seastar should not breask this assumption.

This commit keeps it around by adding a boolean
`_aborted` member to subscription used to make
sure the callback is called at most once,
but otherwise, the function is not destroyed
until the whole sunscription is destroyed.

Added unit test to check that.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>

Closes scylladb/seastar#2360