remove vestigial OSv support
Per #2392 this support is obsolete and can be removed.

This change removes all the code in the HAVE_OSV blocks, the README
associated with OSV, updates one comment listing osv as one of the
backends, and removes an asm linker directive which was OSv specific.

It also trims some spurious trailing whitespace in the modified files.

Fixes #2392.